fix some warnings in mouse controller

+++ src/core/MouseController.cpp
@@ -18,15 +18,16 @@ void MouseController::MouseCallback(GLFWwindow* window, double xpos, double ypos
     if (!activeCamera) return;
 
     if (firstMouse) {
-        lastX = xpos;
-        lastY = ypos;
+        lastX = static_cast<float>(xpos);
+        lastY = static_cast<float>(ypos);
         firstMouse = false;
     }
 
-    float xoffset = xpos - lastX;
-    float yoffset = lastY - ypos;
-    lastX = xpos;
-    lastY = ypos;
+    float xoffset = static_cast<float>(xpos) - lastX;
+    float yoffset = lastY - static_cast<float>(ypos);
+
+    lastX = static_cast<float>(xpos);
+    lastY = static_cast<float>(ypos);
 
     activeCamera->ProcessMouseMovement(xoffset, yoffset);
 }
